Update tests to accomodate runner changes.
